Add `MinimalGenerics` and allow users to set a `GenericStrategy`
