+++ src/com/esotericsoftware/kryo/util/DefaultGenerics.java
@@ -26,74 +26,18 @@ import java.lang.reflect.TypeVariable;
 
 /** Stores the generic type arguments and actual classes for type variables in the current location in the object graph.
  * @author Nathan Sweet */
-public final class DefaultGenerics implements Generics {
-	private final Kryo kryo;
-
-	private int genericTypesSize;
-	private GenericType[] genericTypes = new GenericType[16];
-	private int[] depths = new int[16];
+public final class DefaultGenerics extends BaseGenerics {
 
 	private int argumentsSize;
 	private Type[] arguments = new Type[16];
 
 	public DefaultGenerics (Kryo kryo) {
-		this.kryo = kryo;
-	}
-
-	@Override
-	public void pushGenericType (GenericType fieldType) {
-		// Ensure genericTypes and depths capacity.
-		int size = genericTypesSize;
-		if (size + 1 == genericTypes.length) {
-			GenericType[] genericTypesNew = new GenericType[genericTypes.length << 1];
-			System.arraycopy(genericTypes, 0, genericTypesNew, 0, size);
-			genericTypes = genericTypesNew;
-			int[] depthsNew = new int[depths.length << 1];
-			System.arraycopy(depths, 0, depthsNew, 0, size);
-			depths = depthsNew;
-		}
-
-		genericTypesSize = size + 1;
-		genericTypes[size] = fieldType;
-		depths[size] = kryo.getDepth();
-	}
-
-	@Override
-	public void popGenericType () {
-		int size = genericTypesSize;
-		if (size == 0) return;
-		size--;
-		if (depths[size] < kryo.getDepth()) return;
-		genericTypes[size] = null;
-		genericTypesSize = size;
-	}
-
-	@Override
-	public GenericType[] nextGenericTypes () {
-		int index = genericTypesSize;
-		if (index > 0) {
-			index--;
-			GenericType genericType = genericTypes[index];
-			if (genericType.arguments == null) return null;
-			// The depth must match to prevent the types being wrong if a serializer doesn't call nextGenericTypes.
-			if (depths[index] == kryo.getDepth() - 1) {
-				pushGenericType(genericType.arguments[genericType.arguments.length - 1]);
-				return genericType.arguments;
-			}
-		}
-		return null;
-	}
-
-	@Override
-	public Class nextGenericClass () {
-		GenericType[] arguments = nextGenericTypes();
-		if (arguments == null) return null;
-		return arguments[0].resolve(this);
+		super(kryo);
 	}
 
 	@Override
 	public int pushTypeVariables (GenericsHierarchy hierarchy, GenericType[] args) {
-		// Do not store type variables if hierarchy is empty or we do not have arguments for all root parameters.
+		// Do not store type variables if hierarchy is empty, or we do not have arguments for all root parameters.
 		if (hierarchy.total == 0 || hierarchy.rootTotal > args.length) return 0;
 
 		int startSize = this.argumentsSize;
@@ -143,11 +87,6 @@ public final class DefaultGenerics implements Generics {
 			if (arg == typeVariable || arg.equals(typeVariable)) return (Class)arguments[i + 1];
 		}
 		return null;
-	}
-
-	@Override
-	public int getGenericTypesSize () {
-		return genericTypesSize;
 	}
 
 	public String toString () {
